Drop V2: `Heading` component moved to `UI`
