+++ packages/web/app/src/components/organization/billing/BillingPaymentMethod.tsx
@@ -3,7 +3,8 @@ import clsx from 'clsx';
 import { useMutation } from 'urql';
 import { Section } from '@/components/common';
 import { Button } from '@/components/ui/button';
-import { Heading, Link } from '@/components/v2';
+import { Heading } from '@/components/ui/heading';
+import { Link } from '@/components/v2';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { BillingPlanType } from '@/gql/graphql';
 import { ExternalLinkIcon } from '@radix-ui/react-icons';
