+++ packages/web/app/src/components/project/alerts/create-channel.tsx
@@ -3,7 +3,8 @@ import { useFormik } from 'formik';
 import { useMutation } from 'urql';
 import * as Yup from 'yup';
 import { Button } from '@/components/ui/button';
-import { Heading, Input, Modal, Select, Tag } from '@/components/v2';
+import { Heading } from '@/components/ui/heading';
+import { Input, Modal, Select, Tag } from '@/components/v2';
 import { graphql } from '@/gql';
 import { AlertChannelType } from '@/gql/graphql';
