+++ packages/web/app/src/components/project/settings/model-migration.tsx
@@ -1,7 +1,8 @@
 import { ReactElement, ReactNode, useCallback } from 'react';
 import { useMutation } from 'urql';
 import { Button } from '@/components/ui/button';
-import { Card, Heading, Tooltip } from '@/components/v2';
+import { Heading } from '@/components/ui/heading';
+import { Card, Tooltip } from '@/components/v2';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { RegistryModel } from '@/gql/graphql';
 import { useNotifications } from '@/lib/hooks';
