+++ packages/web/app/src/components/v2/modals/create-access-token.tsx
@@ -4,7 +4,8 @@ import { useMutation, useQuery } from 'urql';
 import * as Yup from 'yup';
 import { PermissionScopeItem, usePermissionsManager } from '@/components/organization/Permissions';
 import { Button } from '@/components/ui/button';
-import { Accordion, CopyValue, Heading, Input, Modal, Tag } from '@/components/v2';
+import { Heading } from '@/components/ui/heading';
+import { Accordion, CopyValue, Input, Modal, Tag } from '@/components/v2';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { TargetAccessScope } from '@/gql/graphql';
 import { RegistryAccessScope } from '@/lib/access/common';
