+++ packages/web/app/src/components/v2/modals/leave-organization.tsx
@@ -3,7 +3,8 @@ import cookies from 'js-cookie';
 import { FaUsersSlash } from 'react-icons/fa';
 import { useMutation } from 'urql';
 import { Button } from '@/components/ui/button';
-import { Heading, Modal } from '@/components/v2';
+import { Heading } from '@/components/ui/heading';
+import { Modal } from '@/components/v2';
 import { LAST_VISITED_ORG_KEY } from '@/constants';
 import { graphql } from '@/gql';
 import { useNotifications } from '@/lib/hooks/use-notifications';
