+++ packages/web/app/src/components/v2/modals/transfer-organization-ownership.tsx
@@ -4,7 +4,8 @@ import { useFormik } from 'formik';
 import { useMutation, useQuery } from 'urql';
 import * as Yup from 'yup';
 import { Button } from '@/components/ui/button';
-import { Heading, Input, Modal } from '@/components/v2';
+import { Heading } from '@/components/ui/heading';
+import { Input, Modal } from '@/components/v2';
 import { ArrowDownIcon, CheckIcon } from '@/components/v2/icon';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { useNotifications } from '@/lib/hooks';
