+++ packages/web/app/src/pages/organization-subscription.tsx
@@ -10,11 +10,11 @@ import { InvoicesList } from '@/components/organization/billing/InvoicesList';
 import { RenderIfStripeAvailable } from '@/components/organization/stripe';
 import { OrganizationUsageEstimationView } from '@/components/organization/Usage';
 import { Button } from '@/components/ui/button';
+import { Heading } from '@/components/ui/heading';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { Card } from '@/components/v2/card';
-import { Heading } from '@/components/v2/heading';
 import Stat from '@/components/v2/stat';
 import { graphql, useFragment } from '@/gql';
 import { OrganizationAccessScope, useOrganizationAccess } from '@/lib/access/organization';
